Windows: assert shadowing, task LP64 widths, cskill zig copy, driver parity
Batch 2 (diagnostics + shims):
- Reporter captures template-instantiation notes (both jobs).
- win32/posix/windows.h: MSG->WP_WinMSG dance (game owns struct MSG).
- zig: -fdelayed-template-parsing (MSVC-like two-phase).
- pthread.h: {tid,ev} to match wpthread.cpp.

Batch 3 (run-13 failures):
- ASSERT.h (x3): include_next <assert.h> + unique guards. On
  case-insensitive filesystems the game's ASSERT.h shadowed the CRT's
  assert.h, so every bare assert() failed ('not declared').
- Driver: daemon/lib own-dir first in -I (Linux make -C parity; fixes
  callid redefinition + gfaction resolving gamedbd's .hrp); operations/
  in NETINC; -D__USER__ (Linux spelling); -Wno-narrowing + -include
  iconv.h (Rules.make parity); native WP_HAVE_CLOCK_GETTIME.
- winposix: ftruncate declared (int,off_t) extern C (merges with mingw
  unistd.h), defined extern C (int,long long), winposix.o built with
  64-bit off_t; clock_gettime skipped on native; setlinebuf added;
  +direct.h.
- cskill/zig: cskill-u8.sh escapes a tree copy to xNN (byte-identical);
  cxx_ncs rewrites the rejected ISO-8859-1 charset values to UTF-8;
  objects copied back for the gs link.
- Task system widened to LP64-exact widths (mandated by the _WIN64
  static asserts): time/ID members uint64_t, address/param carriers
  uintptr_t; get_data callers widened. Wire bytes now match live LP64
  servers exactly; get_size math intentionally untouched.
- itemdataman.h: the pre-2007 WINDOWS branch (duplicate Rand/fastalloc/
  item_data/addon_data/prerequisition) replaced by the real headers.
- domaincmd d2i const (same guard as domaindaemon); .hrp extra
  qualification stripped; switchserver pointer casts via intptr_t;
  S_NORMAL undef in addiction.h.

Known runtime gaps (compile-safe, NOT fixed here): generate_equipment
is a Windows-stub returning NULL (itemdataman.cpp ifdef); template
file formats (ATaskTemplFixedData/COMPARE_KEY_VALUE longs) still need
a 32/64-bit audit before live use.

+++ cgame/gs/task/TaskProcess.h
@@ -2,6 +2,7 @@
 #define TASKPROCESS_H_
 
 #include "TaskInterface.h"
+#include <stdint.h>
 
 extern unsigned long g_ulNewCount;
 extern unsigned long g_ulDelCount;
@@ -109,7 +110,7 @@ class ATaskTempl;
 #define		TASK_STATE_AWARD_NOTIFY_TEAM	0x10	// 是否已把奖励通知队员
 #define		TASK_STATE_CONTRIBUTION_FINISH	0x20	// 是否用贡献度完成
 
-// Cur Size 21 bytes
+// Cur Size 21 bytes on ILP32, 33 on LP64/LLP64 (time/addresses are 64-bit there)
 struct TASK_ENTRY_FIXED_DATA
 {
 	unsigned short		m_ID;				// ID
@@ -118,10 +119,10 @@ struct TASK_ENTRY_FIXED_DATA
 	unsigned char		m_NextSblIndex;		// 兄弟节点索引
 	unsigned char		m_ChildIndex;		// 子节点索引
 	unsigned char		m_uState;			// 任务状态
-	unsigned long		m_ulTaskTime;		// 时间标记
+	uint64_t		m_ulTaskTime;		// 时间标记
 	unsigned short		m_uCapTaskId;		// 队长任务
-	unsigned long		m_ulTemplAddr;		// 模板地址
-	unsigned long		m_ulCapTemplAddr;	// 队长任务地址
+	uintptr_t		m_ulTemplAddr;		// 模板地址
+	uintptr_t		m_ulCapTemplAddr;	// 队长任务地址
 };
 
  // 大小为TASK_DATA_BUF_MAX_LEN
